refactor code and tests (again)
