+++ src/vault/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Vault handlers."""
