+++ tox.ini
@@ -62,7 +62,7 @@ commands =
     isort --check-only --diff {[vars]all_path} 
     black --check --diff {[vars]all_path} 
     mypy {[vars]all_path} --ignore-missing-imports --install-types --non-interactive --explicit-package-bases
-    pylint {[vars]all_path} --disable=E0401,W1203,W0613,W0718,R0903,W1514,C0301,C0103,C0104,R0914,R0915,R0801,W0212,W0719,W0511,R0913,R0912
+    pylint {[vars]all_path} --disable=E0401,W1203,W0613,W0718,R0903,W1514,C0301,C0103,C0104,R0914,R0915,R0801,W0212,W0719,W0511,R0913,R0912,R0902
 
 [testenv:unit]
 description = Run unit tests
